[Container] Move to the core
