+++ docs/src/modules/components/AppContent.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
 import { withStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 
 const styles = theme => ({
   root: {
