+++ docs/src/modules/components/HomeBackers.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { withStyles } from '@material-ui/core/styles';
 import NoSsr from '@material-ui/core/NoSsr';
 import MarkdownElement from 'docs/src/modules/components/MarkdownElement';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 import mapTranslations from 'docs/src/modules/utils/mapTranslations';
 import compose from 'docs/src/modules/utils/compose';
