+++ docs/src/modules/components/HomeFooter.js
@@ -5,7 +5,7 @@ import Interpolate from '@trendmicro/react-interpolate';
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import Grid from '@material-ui/core/Grid';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 import Link from 'docs/src/modules/components/Link';
 import compose from 'docs/src/modules/utils/compose';
