+++ docs/src/pages/getting-started/page-layout-examples/album/Album.js
@@ -12,7 +12,7 @@ import Grid from '@material-ui/core/Grid';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import { withStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 
 const styles = theme => ({
   icon: {
