+++ docs/src/pages/getting-started/page-layout-examples/blog/Blog.js
@@ -16,7 +16,7 @@ import Hidden from '@material-ui/core/Hidden';
 import Link from '@material-ui/core/Link';
 import Button from '@material-ui/core/Button';
 import Divider from '@material-ui/core/Divider';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 import Markdown from './Markdown';
 import post1 from './blog-post.1.md';
 import post2 from './blog-post.2.md';
