+++ docs/src/pages/getting-started/page-layout-examples/pricing/Pricing.js
@@ -13,7 +13,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Link from '@material-ui/core/Link';
 import { withStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 
 const styles = theme => ({
   '@global': {
