+++ docs/src/pages/getting-started/page-layout-examples/sticky-footer/StickyFooter.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Typography from '@material-ui/core/Typography';
 import { withStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 
 const styles = theme => ({
   root: {
