+++ docs/src/pages/layout/basics/basics.md
@@ -7,11 +7,12 @@
 [Responsive layouts](https://material.io/design/layout/responsive-layout-grid.html) in Material Design adapt to any possible screen size.
 We provide the following helpers to make the UI responsive:
 
-- [Grid](/layout/grid/): The grid creates visual consistency between layouts while allowing flexibility across a wide variety of designs.
-- [Breakpoints](/layout/breakpoints/): We provide a low-level API that enables the use of breakpoints in a wide variety of contexts.
-- [useMediaQuery](/layout/use-media-query/): This is a CSS media query hook for React.
+- [Grid](/layout/grid/): The Material Design responsive layout grid adapts to screen size and orientation, ensuring consistency across layouts.
+- [Layout](/layout/layout/): Components and options for laying out your project.
+- [Breakpoints](/layout/breakpoints/): API that enables the use of breakpoints in a wide variety of contexts.
+- [useMediaQuery](/layout/use-media-query/): This is a CSS media query hook for React. It listens for matches to a CSS media query.
 - [Hidden](/layout/hidden/):
-The hidden component can be used to change the visibility of the elements.
+Quickly and responsively toggle the visibility value of components and more with our hidden utilities.
 
 ## z-index
