+++ docs/src/pages/layout/layout/FixedContainer.js
@@ -1,16 +1,14 @@
 import React from 'react';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Typography from '@material-ui/core/Typography';
-import Container from '@material-ui/lab/Container';
+import Container from '@material-ui/core/Container';
 
 function FixedContainer() {
   return (
     <React.Fragment>
       <CssBaseline />
       <Container fixed>
-        <Typography component="div" style={{ backgroundColor: '#fff', height: '100vh' }}>
-          Fixed
-        </Typography>
+        <Typography component="div" style={{ backgroundColor: '#cfe8fc', height: '100vh' }} />
       </Container>
     </React.Fragment>
   );
