+++ packages/material-ui/src/Container/Container.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
-import { withStyles } from '@material-ui/core/styles';
-import { capitalize } from '@material-ui/core/utils';
+import { withStyles } from '../styles';
+import { capitalize } from '../utils';
 
 export const styles = theme => ({
   root: {
