+++ packages/material-ui/src/Container/index.d.ts
@@ -1,1 +1,2 @@
 export { default } from './Container';
+export * from './Container';
